+++ example/package.json
@@ -38,10 +38,8 @@
     "react-native-fast-image": "^8.3.3",
     "react-native-fbsdk": "^3.0.0",
     "react-native-gesture-handler": "^1.8.0",
-    "react-native-gifted-chat": "^0.16.3",
     "react-native-google-places-autocomplete": "^2.1.3",
     "react-native-image-picker": "^3.2.1",
-    "react-native-lightbox": "^0.8.1",
     "react-native-linear-gradient": "^2.5.6",
     "react-native-loading-spinner-overlay": "^2.0.0",
     "react-native-maps": "^0.27.1",
